Implement keys method
This is more efficient than using the current iter method and
mapping to the keys since the iter method produces the
keys multiple times (once for each value). Granted, this isn't
going to be a huge problem with query strings, but query-map
is also used for stage-variables and path-parameters in AWS
implementations ... Aaand granted that it's ALSO not a huge
problem there because they are restricted to single values -
but for more general uses of query-map, this may be a boon.